# Conflicts:
#	.idea/workspace.xml

+++ src/Controller/BlogController.php
@@ -47,7 +47,7 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/blog/posts", name="posts_show")
+     * @Route("/posts", name="posts_show")
      * @return Response
      */
     public function blogPosts(): Response
@@ -68,7 +68,7 @@ class BlogController extends AbstractController
 
 
     /**
-     * @Route("/blog/post/{url_alias}", name="post_show")
+     * @Route("/post/{url_alias}", name="post_show")
      * @param $url_alias
      * @return Response
      */
@@ -104,12 +104,13 @@ class BlogController extends AbstractController
         return $this->render('post/showPost.html.twig',[
                'post' => $post,
             'latests' => $latests,
+            'user' => $user,
             'form' => $form->createView()
         ]);
 
         }
     /**
-     * @Route("/blog/{username}/posts", name="user_posts")
+     * @Route("/{username}/posts", name="user_posts")
      * @param $user
      * @return Response
      */
@@ -132,14 +133,14 @@ class BlogController extends AbstractController
     /**
      * @Route("/profile/{username}", name="user_profile")
      * @param $user
-     * @return Response
      * @IsGranted("ROLE_USER")
+     * @return Response
      */
     public function renderProfile(User $user): Response
     {
         $connecteduser = $this->getUser();
         if($connecteduser->getUsername() != $user->getUsername() || $connecteduser == null){
-            throw $this->createAccessDeniedException();
+            return $this->render('security/error.html.twig');
         }
         $posts = $this->getDoctrine()
             ->getRepository(Post::class)
@@ -200,6 +201,12 @@ class BlogController extends AbstractController
     {
         $slugify = new Slugify();
         $post = $this->getDoctrine()->getRepository(Post::class)->findOneBy(['id' => $id]);
+        $userConnected = $this->getUser()->getUsername();
+        $userr = $this->getDoctrine()->getRepository(User::class)->findOneBy(['id'=> $post->getUser()->getId()]);
+        if(strcmp($userConnected , $userr->getUsername()) !== 0){
+            return $this->render('security/error.html.twig');
+        }
+
         $form = $this->createForm(UpdateType::class,$post);
         $user = $this->getUser();
 
@@ -226,41 +233,46 @@ class BlogController extends AbstractController
     /**
      * @Route("/posts/delete-post/{id}", name="delete_post")
      * @param $id
+     * @IsGranted("ROLE_USER")
      * @return Response
      */
     public function renderDelete($id): Response
     {
         $post = $this->getDoctrine()->getRepository(Post::class)->findOneBy(['id' => $id]);
+        $userConnected = $this->getUser()->getUsername();
+        $userr = $this->getDoctrine()->getRepository(User::class)->findOneBy(['id'=> $post->getUser()->getId()]);
+        //dd($userConnected,$userr);
+        if(strcmp($userConnected , $userr->getUsername()) !== 0){
+            return $this->render('security/error.html.twig');
+        }
         $em = $this->getDoctrine()->getManager();
-        $user = $this->getUser();
         $em->remove($post);
         $em->flush();
-        return $this->redirectToRoute('profile',['username'=> $user->getUsername(),'_fragment'=> 'v-pills-post-management']);
+        return $this->redirectToRoute('profile',['username'=> $userConnected,'_fragment'=> 'v-pills-post-management']);
     }
 
     /**
      * @Route("/posts/delete-comment/{id}", name="delete_comment")
      * @param $id
+     * @IsGranted("ROLE_USER")
      * @return Response
      */
     public function deleteComment($id): Response
     {
         $comment = $this->getDoctrine()->getRepository(Comment::class)->findOneBy(['id' => $id]);
-        if($this->getUser()!==null){
+        if(strcmp($this->getUser()->getUsername() , $comment->getAuthor()->getUsername()) !== 0){
+                return $this->render('security/error.html.twig');
+        }
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($comment);
+            $post = $this->getDoctrine()->getRepository(Post::class)->findOneBy(['id' => $comment->getPost()]);
+            $em->flush();
+            return $this->redirectToRoute('post_show', ['url_alias' => $post->geturl_alias()]);
 
-            if($this->getUser()->getUsername()!=$comment->getAuthor()->getUsername()){
-                $this->createAccessDeniedException();
-            }
-    }
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($comment);
-        $post = $this->getDoctrine()->getRepository(Post::class)->findOneBy(['id'=>$comment->getPost()]);
-        $em->flush();
-        return $this->redirectToRoute('post_show',['url_alias' => $post->geturl_alias()]);
     }
 
     /**
-     * @Route("/blog/post/{id}", name="showsPost")
+     * @Route("/post/{id}", name="showsPost")
      * @param $id
      * @return Response
      */
